Reorganize state(parameter) code
