+++ src/webgl/api/README.md
@@ -1,0 +1,3 @@
+# webgl/api
+
+This directory contains "metadata" about the webgl api, lists of constants, parameters, settings etc that allow luma.gl to expose webgl functionality in a better way.
