+++ src/webgl/texture.js
@@ -1,7 +1,8 @@
 /* eslint-disable no-inline-comments, max-len */
 import GL from './gl-constants';
 import {WebGLBuffer} from './api';
-import {withParameters, isWebGL, ERR_WEBGL, isWebGL2, ERR_WEBGL2} from './context';
+import {isWebGL, ERR_WEBGL, isWebGL2, ERR_WEBGL2} from './context';
+import {withParameters} from './context-state';
 import Resource from './resource';
 import Buffer from './buffer';
 import {uid} from '../utils';
